[TSAR-cmake] Update scripts for CMake.
It is possible to build Clang, opt and TSAR in the same solution and install them in the specified folder.
Patches for LLVM and Clang have been added.
There are two types of patches. The first one has the tsar prefix and is necessary to build TSAR. 
The second one eliminates errors in LLVM and Clang.

git-svn-id: http://svn.dvm-system.org/svn/dvmhrepo@3102 c90b6321-6eb1-4274-a4bb-7129d0cd3ae9