[TSAR] Use results of alias analysis phases in privatizability analysis.
Note. The presence of the operation of taking the address &X does not mean that 
X can not be privatized. This means that the address of the original variable X
must be available. For example, a set of private variables X1, ..., XN (where
N - number of processors) can be used instead of accesses to X, and X can be
marked as a shared variable.

git-svn-id: http://svn.dvm-system.org/svn/dvmhrepo@3217 c90b6321-6eb1-4274-a4bb-7129d0cd3ae9