[TSAR] Fix analysis in case when location access is performed by pointer.
Location can not be declared as copy in or copy out without
additional analysis because we do not know which memory must
be copy. Let see an example:
for (...) { P = &X; *P = ...; P = &Y; } after loop P = &Y, not &X.
P = &Y; for (...) { *P = ...; P = &X; } before loop P = &Y, not &X.
Note that case when location is shared, but pointer is not shared
may be dificulty to implement for distributed memory, for example:
for(...) { P = ...; ... = *P; } It is not evident which memory
should be copy to each processor.

git-svn-id: http://svn.dvm-system.org/svn/dvmhrepo@3241 c90b6321-6eb1-4274-a4bb-7129d0cd3ae9