[TSAR] Fix, two conditions has been added to check whether whole base location is written in a loop.
Let us propose the following explanation. Consider a loop where some location Loc is written and this memory is going to be read after an exit from this loop. It is possible that base for this location BaseLoc covers this location, so not a whole memory comprises BaseLoc is written in the loop. To avoid a loss of data stored before the loop execution in a part of memory which is not written after copy out from this loop the BaseLoc must be also treated as a first private.

git-svn-id: http://svn.dvm-system.org/svn/dvmhrepo@3622 c90b6321-6eb1-4274-a4bb-7129d0cd3ae9