[TSAR] Implement provider pass to transfer results from function passes to a module pass.
There are different passes for different entities: module, function, etc. It is not easy to transfer results of analysis between passes of different levels.

For example, to obtain results of function analyzes from a module pass the specified function passes will be executed on the fly for the specified function many times. Other problem is that a default constructor is used to create this function passes in the pass manager so it is not possible to specify additional parameters to initialize a such passes.

To avoid this problems a provider pass could be used.

git-svn-id: http://svn.dvm-system.org/svn/dvmhrepo@3718 c90b6321-6eb1-4274-a4bb-7129d0cd3ae9