[TSAR] Enable to merge multiple sources before analysis.
Some new options have been added.
At this moment clang::ASTMergeAction is used, however there is some errors.
For example if declaration of void f(int) is imported before definition, parameters will not be imported correctly.
In this case it is not possible to generate LLVM IR (assertion fail).

Some improvements also should be made.
For example, if a name of a static function already exists it should be changed before two AST will be merged.