[TSAR] Update memory matcher to simplify access to its result from function passes.
A function pass manager can not run module passes. This manager is used
to access results of functions passes from a module pass (A). These
function passes are run on the fly. If some of these passes need to access
results of other module pass these results should be obtained before
a module pass A. Then results should be stored and transfered into the
inner pass manager which runs function passes on the fly.

Pass provider and analysis wrapper pass may be used to implement this.