[TSAR, Unparser] Fix unparsing of void * array.
void foo(void *A, int N) {
  int *B = A;
  B[0] = 1;
}
In this case estimate memory location will be build for &A.
The type of A is a pointer to void. It means that base
type of DIDerivedType of A will be null.

More complicated cases are not considered since cast operations
are not processed.