[TSAR, AliasTree] Fix splicing and merging of estimate memory trees.
splicePrev() and mergePrev() is identical now.
Some links should not be broken.
Children of a node should not become children of a previously splicing (merging) node.
This is important in case of updating staring point of memory chain.

There was the following bug.
 <*A,?> <*A,?>->splicePrev(<*A, 8>)
    |
 <A[1], 8>
The result was
 <*A, ?>
    |
 <*A, 8>
    |
 <A[1], 8>

Instead of
     <*A, ?>
     /     \
 <*A, 8> <A[1], 8>

Note, that in this case <*A, 8> and <*A, ?> is a part of the same chain (see next and prev methods).

Some forgotten links should be broken (remove nodes from children list).