[TSAR, Instr] Some fix to evaluate inline functions.
The 'linkonce_odr' or 'linkonce' linkage will be set for an 'inline'
function that will not be inlined by Clang. If there are many versions
of this function in a source code then only one will be used.
We change these linkages to 'internal' so a function that has been
instrumented will be used in a instrumented file.