[TSAR, Inline] Intermediate, add one-step checkers for calls.
We change approach to constraction of forward declaration set for a function.
Now, AST is used before conservative seach which is base on relexing of the function definition.

New checkers for calls have been added.
However, at this moment we do not update forward declarations after inlining.

So, implementation of inlining of chain of calls is not completed yet.