[TSAR, Inline] Use set of all identifiers to create the new one without conflicts.
These identifiers is used to prevent conflicts when new identifiers
are added to the source code. It is convenient to avoid intersection with
all available identifiers (including the local ones). For example,
if chain of calls should be inlnined in a function, it is necessary to
check that all new identifiers do not hide forward declarations of all
functions in this chain.