[TSAR, Inline] Update check which identifies whether addition braces is necessary.
We previously explored whether braces which surrounds inline call should
be omitted. The braces should be omitted in case of calls inside
a declaration which is used outside of declaration statement.
For example, int x = f(); ... x ... ;

However, to implement this check all references to declarations should be
collected accurately. Note, that there are may be not only variable
declaration, but for example type declarations.

struct A { int X; } A1 = { .X = f() };
struct A A2;

Declaration of structure A should be visible after the first declaration
statement. However, the previous implementation of discussed check did
not consider this case and it inserts braces.
So, the code become uncompilable.

To solve this issue we add braces only if they are necessary.
If statement with call is not inside a compound statement braces should be
added after inlining: if(...) f(); -> if (...) { /* inlined f() */ }