[TSAR, Inline] Fix, use type visitor to find external type declarations.
In some cases Visit...TypeLoc methods do not visit some types.
For example, TypeLocs will be ommited for parameters of a function which has been imported.
However, usage of Visit...Type methods solves this problem.

The second improvement is that we use the first redeclaration of external
dependencies to remember this dependence.