[TSAR] Avoid use of -O0 which leads to optnone.
Behaviour of Clang has been changed if -O0 is used (if no options is
used then -O0 use used by default). In this case Clang set 'optnone' attribute
for each function. This attribute prevents optimisations, so LLVM passes
do nothing. It is possible to use -Xclang -disable-O0-optnone. However,
this is not recomended because -O0 has a special meaning. It means
disable optimizations. For example, it also marks all functions as
'noinline'.

So, we use: -O1 -Xclang -disable-llvm-passes