[TSAR, Tfm, Inline] Fix, insert #pragma 'nomacro' before function body.
Bounds should be specified for 'nomacro' directive.
Rounded braces for the body of an inlined function are always exist.
However, in case of non-void return value there is a declaration of return value before the inlned body.
So, we make this fix and move 'nomacro' directive immediately before the body.