[TSAR, Tfm, Inline] Fix processing of statements with multiple calls.
(1) We should preserve the order of functions execution after inlining.
So, we use std::vector to store calls and llvm::DenseMap to perform search of a call.

(2) It is not safe to replace a statement multiple times
if the length of replacement differs from the length of the original statement.
So, we use InsertAfter and InsertBefore methods instead.