[TSAR, Tfm, Inline] Fix, use expansion range to perform transformation.
(1) In case of LocalLexer Lex(...) it is necessary to use expansion range, because getSourceRane() may returns range with bounds in different buffers.
So, it is not possible to relex such range. For example, `MACRO(x) f()`, where `MACRO` is `x +`.

(2) InsertTextAfter and InsertTextBefore should use expansion locations because in the example mentioned above file location points to x.
So, the body of `f` will be inserted between `MACRO` and `(x)`