[TSAR, Tfm, Clang] Fix, out of range.
The most of rewriter methods treat SourceRange as a range of tokens (the last location is a position of the last token).
So, rewriter adds size of the last token to the offset of the last position.
This leads to out of range error.
To overcome this issue we use rewriter buffer to obtain correct size of a file that should be transformed.