[TSAR, Tfm, Clang] Major update of ExternalRewriter.
Now, we use [begin, end] range to represent each location after transformation:
- 'begin' points to the beginning of a string inserted before the original location;
- 'end' points to the position of the original location after transformation.

The previous implementation has used only 'end' location, this leads to incorrect transformation in some cases.
For example,
  ... ; <token>
At first, we insert some text before token.
  ... ; <text> <token>
And now, we want to remove ';'.
If only 'end' is available for original <token> location then it is not possible to determine whether <text> is a part of removed range or not.
Note, that <text> may be a part of transformed original range of ';'. In this case we should remove '; <text>'.
However, if <text> is a result of <token> transformation we should remove ';' only.

If we use 'begin' location then it is possible to distinguish these cases.