[TSAR, Tfm, Clang] Fix, replace text works incorrect in the case of a whole file.
It seems that there is error in implementation of Buffer.ReplaceText().
This error does not arise if a very small file is transformed.
In the case of large files the root 'RopePieceBTreeNode' is not a leaf.
However, if the whole body of a file will be removed before insertion getNumChildren() for the root returns 0.
So, the line 513	in RewriterRoop.cpp (llvm-700) produces an error (i = e - 1 => i = 0 - 1, i - unsigned).