[TSAR, AliasTree] Fix, update, DW_OP_LLVM_fragment with size 0 means that size of location is unknown.
void foo(int *Y);
void bar() {
  int X[] = { 1, 2 };
  foo(X);
}
Two meta-locations should be created <X, 12> and <X, ?>.
The previous version of analyzer has created location with empty expression in both cases.
This leads to assertion fail. Now, in the second case the expression is {DW_OP_LLVM_fragment 0, 0} and this means that size in unknown.