[TSAR, AliasTree] Fix evaluation of 'inttoptr' casts.
(1) Alias analysis for non-pointer values always return NoAlias.
So, it is not possible to strip 'inttoptr' when base location is constructed.
However, we may attach different values with 'inttoptr' to the same estimate memory location.
For example, the following bases are equal:
  'inttoptr i64 %Y to i64*'
  'inttoptr i64 %Y to %struct.S*'

(2) We use underlining value for 'inttoptr' to build DIUNknonwMemory locations.
We also strips 'load' instructions if it exists.
For example,
  '%3 = load i64, i64* %Y.addr'
  '%4 = inttoptr i64 %3 to i64*'
%4 produce <*Y,?> metadata-level unknown memory location.
After promotion we obtain
  'inttoptr i64 %Y to i64*'
which produces the same <*Y,?> location.