[TSAR, AliasTree] Fix, add locations from a chain with root if necessary.
Consider an example:
  struct S { int X, Y; };

  int foo(struct S In) {
    return In.X + *(int *)(((char*)&In)+2) + In.Y;
  }

IR-level alias tree contains a chain EM1=<In, 8> <- Em2=<In, ?>.
Root of the estimate tree is <In, ?>.
EM1.front() == EM2.front() == Root.front(), however sizes of EM1 and EM2 are different.
So, we should insert <In, 8> in metadata-level alias tree.
Otherwise, after promotion metadata-level alias tree will be incorrect because <In, 8> will be inserted as a promoted fragment.
And this fragment will not alias with other locations (for example, <In, ?>) because there is no corresponding location <In, 8> in the previously builded (before promotion) tree.