[TSAR, Tfm, Propagate] Remove check for operand number for propagated variables.
It seems that a check for variable names is sufficient.
It is also important that a check for operand number prevents propagation in array subscripts and pointer arithmetic.
In LLVM IR it will be GEP with OpNo from 1 to N, however in AST it will be BinaryOperaotr or ArraySubscriptExpr with OpNo from 0 to 1.