[TSAR, AST, Import] Extend information about import process.
Now, we store merged locations separatly for each declaration.
The previouse implementation had the following issue:

  (1) int A, B;
Merge:
  (2) int A, B;
  (3) int B;
Merged locations to getLocStart()
  (1) <- (2) (3)
Merged locations to getLocation()
  (1.A) <- (2)
  (1.B) <- (2) (3)

So, we do not known correspondence between locations for getLocStart() and getLocation() for merged declarations.
However, it is convinient to use getLocStart() to determine whether two global declarations are declaraed in a single statement.
Hence, it is not possible to determine whether statement (2) contains two declarations.