[TSAR, AliasTree] Do not use locations in a source code to check that memory is corrupted.
Consider an example:
void foo(int *A) {
  A[I] = ...
  A[I] = ...
}

Originaly there are two loads from alloca with name A and each of this loads produce a metadata memory location.
These locations are not equal because the contains different locations in a source code (column and line).
However, after SROA alloca and loads will be removed and register %A will be used instead.
Hence if we rebuild metadata locations we obtain identical locations for both operators which do not contain origin column and line.
The new location contain position which points do definition of a variable A.
This means that original locations will be always treated as corrupted locations.

To avoid this behaviour we add base representation for DIMemory: MDNode{ <pointer to base>, <data> }.
For example, <data> may contains list of locations in a source code and
<pointer to base> is an MDNode, which contains variable, expression and some flags.
We use base representation to check whether a memory location is corrupted.