[TSAR, AliasTree] Strip GEP if possible.
In case of sequence of GEPs try to subsequently strip all GEPs to build
a single estimate memory tree with the whole array as a root.
It is not safe to extend GEP memory location if its current size Loc.Ptr
is known and it is greater the size of element because if we do not known
the size of the whole array we can not check that <Loc.Ptr, Loc.Size> is
inside <pointer to array, size of array>. However, we assume that
unknown size can not be greater than allocated size.
Note, if array dimensions have constant size than Size == Loc.Size and
level will be striped.