[TSAR, Memory, Dep] Improve loop invariant check.
Check whether a not promoted memory locations are changed in a loop.
If there is node 'def' instructions then appropriate value is loop invariant.

Let us consider a test shared_11.c.

double U[100][100];
int IStart, IEnd, JStart, JEnd;

void foo() {
  for (int I = IStart; I < IEnd; ++I)
    for (int J = JStart; J < JEnd; ++J)
      U[I][J] = U[I][J] + 1;
}

A variable 'JStart' is loop invariant for a loop 'I'.
However, this variable is global and it can not be promoted to registers.
Hence, 'isLoopInvariant()' method for SCEV expressions is not worked.
To overcome this issue we check abscence of write operations to JStart.