[TSAR, Delinearize] Do not search metadata for 'alloca' if it is treated as a base of array.
'Alloca' could be an array base in case of array of constant size or VLA.
In other cases 'alloca' could be an address of array base only.
However, preliminary collection of array accesses may temporary consider it as a base of array.
For example, void (int (*A)[100]) { ... } produces a store to alloca 'A.addr'.
If we use metadata in this case we will consider 'A.addr' as a base of array with two dimensions.
So, postprocessing of preliminary collected array accesses would not drop 'A.addr' as a not array.
Hence, we obtain two arrays with the same metadata 'A.addr as a base' and 'A.addr as an address of base'.