[TSAR, AliasTree] Fix, add parameters which are not accessed into an alias tree.
We also remove assert from isSameAfterRebuild() because if memory becomes
unused after transformation then it may not be presented in alias tree.

void foo(double *X, int N) {
  for (int I = 1; I < N - 1; ++I) {
      double T[2];
      T[0] = X[I-1];
      T[1] = X[I+1];
  }
}

This function do nothing so it will be transformed to a function with empty body.
However, metadata-level alias tree already contains <*X, 8> (single double) and <*X, ?> (array of doubles).
So, it is necessary to add unused parameter as a pointer to single double and as an array.
Unfortunatly in some case <*X, ?> is redundant and may hide some locations (with known sizes) in dependence analysis results.
Hence, we add only <*X, 8> and remove assert from isSameAfterRebuild().