[TSAR, Memory, Trait, Dep] Add 'redundant' trait.
Now, we determine redundant locationsis which can be removed after some transformations.
Existance of this location may degrade analysis results,
so it may be useful to perform source-to-srouce transformations to remove these locations before analysis.

We also print some traits separately for some locations (along with these traits for the whole alias node).