[TSAR, AliasTree, Metadata] Update, disable some check.
The current commit relates to commit 'Fix in case of &X + 0 locations.' (01.08.2019).
After the mentioned commit 'getelementptr' with zero offset is not a part of alias tree (it is stripped to base pointer).
So, the removed check has prevented insertion of T[0] in alias tree
(in case of 'int T[2]') because T is a pointer to an array and
getelementptr with 0 offset is not a part of alias tree.