[TSAR, Memory] Major update, replace MemoryLocation with MemoryLocationRange in data-flow analysis.
MemoryLocationRange allows to perform more accurate analysis.
For example:
void foo(double *X, int N) {
  double T[2];
  for (int I = 1; I < N - 1; ++I) {
    T[0] = X[I-1];
    T[1] = X[I+1];
    X[I] += T[0] + T[1];
  }
}

If MemoryLocation is used def-use set does not contain <T, 8> (T[0]).
It has <T,16> (the whole array) because addresses of <T,8> and <T,16> are the same.
Memory location <T,16> contains T[0] and T[1].
Hence, it will be updated partially in T[0] =... and T[1] = ....
This means that <T,16> is 'MayDef' not 'Def' and T[0] could not be recognized as private.
So, the whole <T,16> could not be also analyzed successfully.

If we use MemoryLocationRange then def-set will contain:
(1) <T, 0, 8> after processing T[0]
(2) <T, 8, 16> after processing T[1].
Hence, def-use set will contain <T, 16> with 'Def' property instead of 'MayDef'.
And T will be recognized as privitizable.