[TSAR, Memory, Dep] Fix, try to update traits for corrupted locations.
Note, that there was a bug in a previous check for redundant memory.
We should check the whole values bounded to corrupted memory because.
Only if the whole values is redundant then memory is redundant.
We also should check whether  ancestors in estimate memory tree for bounded values.
These ancestors should be also redundant.