[TSAR, Memory, Dep] Fix, major improvment, reimplement combining of traits.
The main reason is that we should combine traits (in case of corrupted and merged location)
before traits for appropriate location would be updated in a pool.
Note, to determine whether it is safe to update trait in pool the original trait and the new one should be known.
This is important in case of last private and second to last private for example (after loop rotate).