[TSAR, AliasTree] Fix, strip element to the whole allocated memory instead of dimension.
int A[10][10];
It is not safe to strip <A[I][J],4> to <A[I],10> even in case of inbounds attribute in GEP.

If the inbounds keyword is present, the result value of the getelementptr
is a poison value if the base pointer is not an in bounds address of an
allocated object, or if any of the addresses that would be formed by
successive addition of the offsets implied by the indices to the base
address with infinitely precise signed arithmetic are not an in bounds
address of that allocated object.

So, it is not safe to strip it to <A,100> only because allocated object
is a whole array.