[TSAR, AliasTree, Metadata] Fix, not only alloca and global could be an address of memory with attached metadata.
Structures are pass by reference in general. So, in case of

  struct STy { int X; double Y; };
  double foo(struct STy S) { ... }

the following LLVM IR will be built

  define dso_local double @foo(%struct.STy* %S) #0 !dbg !8 {
  entry:
    %Res = alloca double, align 8
    %I = alloca i32, align 4
    call void @llvm.dbg.declare(metadata %struct.STy* %S, metadata !18, metadata !DIExpression()), !dbg !22
    ...
  }

DIExpression(DW_OP_LLVM_fragment, 4, 8) should be used to represent S.Y.
So, we should know that %S is a pointer to a structure (like an alloca),
otherwise buildDIExpression() add DW_OP_Deref at the beginning of this expression.