[TSAR, Memory] Fix usage of MemoryLocationRange in MemorySet.
There was a bug in previous implementation of tsar::MemorySet.
If we add into a set <A,0,4> and <A,8,12> we obtain <A,0,12>.
MemoryLocationRange does not allow to store range with holes,
so now we store list of location ranges with the same pointer and
the mention above case are correctly processed.