[TSAR, Pass, APC] Allow to specify necessary passes which should be executed before a current one.
This oportunity may be useful if there is no explicit dependence between
two passes, however results of the necessary pass drastically influence on
the results of the current one. So, there are no reasons to execute
the current pass if necessary passes has not been executed.

For example, APC passes should be executed in a predefined order at
different analysis steps. Hence, it is not possible to use built-in
capabilities to specify dependencies of the passes. So, if result of
some pass is not requested by the user (via appropriate option) necessary
passes could be ommited. Note, that some of these passes may emit some
warnings, which are not useful in this case and also should be ommited.