[TSAR, APC, Array] Use alias tree to find raw representation of an array.
In some cases 'getRawDIMemoryIfExists(Ctx, Loc)' could not find appropriate
representation because DIAliasTree may contains derived representation of an array
which is not equal to Loc. For example, it may contains additional source code locations
or it may contains expression differs from expression in Loc.
In this case raw representation would not be found.
So, we search appropriate memory in alias tree and use this memory to obtain raw representation.