[TSAR, IPO, Attribute] Add sapfor.direct-user-callee attribute.
This attribute means that user-defined functions which are called from
a specified one (may be transitively) can be called directely only.

It is possible to use this attribute to not insert parallel directives
inside inderect recursion. For example, dynamic analysis may determine
absence of dependencies in a loop. However, this loop may indirectly
calls its parent function. So, isertion of parallel directives before
this loop may lead to incorrect program execution.