[TSAR, Utils, Metadata] Fix search for metadata attached to a global variable.
DIGlobalVariableExpression may contain not empty expression.
This expression may specify offset from IR-level pointer to the global object.
For example, in case of Fortran all arrays (inside a function) with constant sizes will be stored in a common global pool.
In this case expression specifies offset from the beginnig of the pool.
Hence, this expression should not be stored in DIMemoryLocation in this case
because DIMemoryLocation::Expr represent offset from a DIMemoryLocation::Var instead of an offset from some LLVM-level address.