[TSAR, Parallel, OpenMP] Do not ignore redundant locations which allow parallelization.
If -fignore-redundant-memory option is set then redundant locations should be ignored.
However, if traits do not specified for some of source-level variables it
is not possible to create 'parallel' directive. So, we do not ignore
redundant memory if its traits do not prevent parallelization this allow
us to build parallel program even in case of -fignore-redundant-memory option.