[TSAR, Delinearize] Fix, strip casts from sub-range address instruction.
We add a check to ensure that GEPs are used to compute pointer to sub-range.
For example, if arithmetic operations followed by int-to-pointer cast
are used to compute address of array element, there are not GEP instructions.
Note, that sometimes instcombine pass inserts bitcast instruction between
an address computation and a memory access instruction (for example it may cast
double* to i64* in case of copy of elements of 'double' type (a[i] = b[j])).
So, we strip cast instructions between memory
access and the first GEP instruction.