[TSAR, Anls, External] Fix, use `ReadOccured` to determine read-only variables.
If results of external analysis do not contain traits for a variable
this does not mean that accesses to this variable do not imply
loop-cariad dependencies. Absance of analysis results only means that
this variable has not been analyzed externally (for example, there was
no corresponding instrumentation, so accesses have been ommited while
dynamic analysis has been performed).