[TSAR, Memory, AliasTree] Fix, reimplement RAUWs invocation while tree is rebuilt.
Previous implementation does not explicitly determine the right order
of replacements. For example, consider replacement of A with C,
and B with C, where A and B are memory locations from the previous
alias tree and C should be inserted into a new alias tree.
The possible orders of occurrences of A and B in the traverse
of alias tree are A then B or B then A.
Assume that raw implementation of B and C is identical.
This means that if raw representation is a key of a memory in the set
of memory uses then it is not clear how to distinguish B from C in such set.

Let we have a set of uses: [A. B].
Assume we meet A firstly and replace it with C. However it is not possible
to insert C in this set because keys for C and B are the same.
The right order is replace B with C at first to obtain [A, C] and
then replace A with C to obtain [C]. Moreover, to identify that we use C
in two replacements it is necessary to mark C as merged
before the second one.

So, we change approach to replacement implementation.
Now we collect all necessary replacements for an alias tree node,
then we build replacement graph and traverse it in a right order.