[TSAR, IPO, Memory, Live] Fix, live set should be accurately propagated though function calls.
In general it is not possible to analyse complex data types accurately.
So it is not safe to initialize analysis with a merged live sets for all function calls.
Instead, we find locations which could be safely analyzed and check whether these locations are live in any functioncall.
For other locations we conservatively assume that they are live.
We also establish relation between actual and formal parameters only if it is possible to ensure its correctness.