[TSAR, AST, Dep] Allow to use variable to represent a memory location even if its size is unknown.
We check that metadata-level memory is larger then allocated
memory. For example:
  - metadata memory location is <X, ?>
  - `int X`
Size of X will be unknown if there is a call which takes &X as
a parameter: bar(&X).
However, in this case it safe to use X to describe <X, ?> because
out of range accesses lead to undefined behaviour.