[TSAR, GUI, CallGraph] Use metadata on client to search function declaration.
Analysis server may change some metadata, especially for builtin functions.
If name is changed, function declaration won't be found.
However, it is possible to use metadata which were stored on client.

We also use linkage name to lookup for function declaration because it
differs from original name in case of C++.