[TSAR, Tool] Add -build-path option to read compile command database.
The addInternalArgs() was also updated to add internal args to the beginning of the list.
Otherwise, these arguments can be implicitly treated as valuses for the last wrong specified user option.

For example if argument for -build-path is missed `tsar ... -p` the first internal argument was treated as a path.