[TSAR, Directive] Fix, skip the last anchor and internal complex expressions.
Let's consider the following nest of expressions:
ZeroOrMore<[
  One<[Identifier]>,
  One<[Identifier]>
]>

For example, iterator 'I' points to the first 'Identifier' and backtrack() is called.
After processing of the first 'One' expression, 'I' will point to its 'EK_Anchor'.
Then 'ZeroOrMore' is processed.

(1) However, 'I' already points to the anchor and 'findAnchor(I)' doesn't change it.
    So, findAnchor(I+1) should be used.
(2) However, in the mentioned example 'I+1' points to the second 'One' expression and
    'findAnchor(I+1)' will points to the end of 'One' expression instead of 'ZeroOrMore'.
    So, while-loop allows us to skip all subsequent complex expressions.