[TSAR, Parallel, Parallel] Add ParallelItem class to represent the result of parallelization.
We also add classes to represent some of the DVMH directives.
We split the parallelization process into two steps.
In the first step, parallel directives will be attached to basic blocks.
This allows us to implement the optimization of a parallel program in further commits.
In the second step, DVMH directives should be inserted into the source code.