[TSAR, Memory, Dep] Search metadata for induction variables in header and latch.
Initially we have only searched metadata for phi-functions in a latch node.
The reason was llvm.dbg.value in header means presence of assignment in the beginning of loop iteration.
For reduction variables such assignment is forbidden and corresponding variables should be marked as private instead of reduction.
However, we may identify variable as induction variable if a corresponding phi-function imlies induction and the variable has not been marked as private yet.

for (int I = 0; I < N; ++I) {
  X = I; //could be considered as induction if it is impossible to prove privatizability
}

This assumption is necessary in case of lower-level inlining.
In the case of such inlining, some metadata in the latch could be lost.