[TSAR, AliasTree] Bound size of memory location if the allocated size is known.
This is a major change.
Now we memory location can have unknown size even if its parent from subsequent chain has known size.
For example foo(&X[5]) yields <GEP(X), ?> because it is not known how 'foo' works.
However, if we know that X has 10 elements, we may build an estimate tree with the following edge <GEP(X), ?> - <X, 10> now.
It simplifies the estimate memory tree and solves some problems with mapping a higher level alias tree to the lower level one.

In the previous implementation if 'clarifyUnknownSize' function is used, the search for a location with unknown size in a lower level alias tree
may decrease the size of this location (if the allocated size is known) and the corresponding node for the location with an unknown size is never found.

We also fixup 'AliasTree::find()'. Now it checks subsequent chains of locations and tries to find the most appropriate location according to the result of the 'stripMemoryLevel()' function.