[TSAR, Anls, Clang] Extract induction bounds from the analysis results.
In APC we use induction bounds to determine whether loop mapping rule should be corrected.

For example, the following loop
  for (int I = 5; I < 100; ++I) { ... A[I] ...}
will be transformed to
  for (int I = 0; I < 95; ++I) { ... A[I+5] ...}
after LLVM IR optimization.
So, the mapping rule is ([I] on A[I+5]) for the transformed loop.
To recalculate the rule ([I] on A[I]) for the initial source code we use the bounds of the induction variable.
If bounds are known then we recalculate the rule.
However, the previous implementation used the canonical loop analysis to determine induction bounds.
Hence, in the case of
   Start = 5;
  for (int I = Start; I < 100; ++I) { ... A[I] ...}
the bounds are unknown but the loop will be transformed in LLVM IR.
To fix this issue, we start using induction variable analysis results to determine loop bounds.
This analysis allows us to determine the necessity of the rule recalculation in this example.