[TSAR, DVMH, Writer] Fix, reimplement fusion of neighbouring directives.
There was a following issue in the previous implementation.
A single directive sometimes has multiple instantiation points
(an instantiation point - is a location in a source code to insert
directive), for example, if this directive is attached to an expression
in a loop head. If we merge neighbouring directives attached to the same
instantiation point we must not affect other instantiation points
of the same directives and we cannot mark the whole directive as skipped.
To overcome this issue we merge directives later when text representations
of neighbouring directives are generated.