[TSAR, Parallel] Allow parallelization of loops that have multiple unreachable exits.
Sometimes LLVM IR for a loop contains `select` instuction which determines loop exits.
This instruction always has default basic block, so, if loop has a single exit in a source code,
this default block will be marked as unreachable.
Hence we can ignore this block and parallelize a corresponding loop if it has only single reachable exit.